# api/openapi.yaml
openapi: 3.0.0
info:
  title: Project Gym API
  version: 2.0.0
  description: |
    API *REST* para gestão de ginásio: planos de treino, exercícios, sessões, avaliações físicas e metas.
    
    
    | Utilizador | *Password* | *API Key* |
    |---|---|---|
    | admin | admin123 | `apikey_admin_001` |
    | treinador | treinador123 | `apikey_treinador_001` |
    | cliente | cliente123 | `apikey_cliente_001` |
servers:
  - url: http://localhost:8080
    description: Project Gym API
...
security:
  - apiKeyAuth: []
  - basicAuth: []
paths:
...
  /users:
    get:
      tags: [Utilizadores]
      summary: Listar utilizadores
      description: |
        **Admin**: Lista todos os utilizadores  
        **Treinador**: Lista clientes do seu âmbito (clientes em planos profissionais do treinador autenticado) e também clientes sem qualquer plano profissional associado a treinador  
        **Cliente**: Não tem acesso
      parameters:
        - $ref: '#/components/parameters/ClienteIdAdminTreinadorQuery'
        - $ref: '#/components/parameters/TreinadorIdAdminQuery'
      responses:
        '200':
          description: Lista de utilizadores
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403': { $ref: '#/components/responses/AcessoNegado' }
  /users/me:
    get:
      tags: [Utilizadores]
      summary: Ver perfil do utilizador autenticado
      responses:
        '200':
          description: Dados do utilizador autenticado
        '401': { $ref: '#/components/responses/NaoAutenticado' }
  /users/me/api-key:
    post:
      tags: [Utilizadores]
      summary: Regenerar *API key*
      responses:
        '200':
          description: Nova API key gerada
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem: { type: string }
                  apiKey:   { type: string }
        '401': { $ref: '#/components/responses/NaoAutenticado' }
  /users/{id}/role:
    put:
      tags: [Utilizadores]
      summary: Alterar role de um utilizador - apenas admin
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [admin, treinador, cliente]
            example:
              role: treinador
      responses:
        '200':
          description: Role atualizado
        '401': { $ref: '#/components/responses/NaoAutenticado' }
        '403':
          description: Sem permissão
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro: { type: string }
              example:
                erro: "Acesso negado. Requer role: admin."
        '404':
          description: Recurso não encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  erro: { type: string }
              example:
                erro: "Utilizador não encontrado."

  # Planos de Treino
...
components:
  parameters:
...
    ClienteIdAdminTreinadorQuery:
      name: cliente_id
      in: query
      required: false
      schema: { type: integer }
      description: |
        Filtrar por cliente (apenas **admin** ou **treinador**).
        Para **treinador**, só funciona para clientes do seu âmbito (clientes em planos profissionais do treinador autenticado)
        e para clientes sem qualquer plano profissional associado a treinador.
    TreinadorIdAdminQuery:
      name: treinador_id
      in: query
      required: false
      schema: { type: integer }
      description: Filtrar clientes por treinador (apenas admin)
...
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    basicAuth:
      type: http
      scheme: basic
  responses:
    NaoAutenticado:
      description: Não autenticado
      content:
        application/json:
          schema:
            type: object
            properties:
              erro: { type: string }
          example:
            erro: "Autenticação necessária. Faça login em /login ou use X-API-Key."
    AcessoNegado:
      description: Sem permissão
      content:
        application/json:
          schema:
            type: object
            properties:
              erro: { type: string }
          example:
            erro: "Acesso negado. Requer role: treinador ou admin."
...